requirements.txt
fastapi
uvicorn
groq
pydantic
supabase
python-dotenv
requests